+++ packages/components/.storybook/preview.js
@@ -1,6 +1,5 @@
 import './storybook.css';
 
 import '../src/assets/tailwind.scss';
-import '../src/assets/index.scss';
 
 export const parameters = {};