+++ packages/components/src/components/forms/Switch/Switch.tsx
@@ -58,7 +58,7 @@ export const Switch = React.forwardRef<ElementType, SwitchProps>(({ className, c
             xl: () => 'h-[38px] w-[76px]',
             xs: () => 'h-[22px] w-[44px]',
           })(scale),
-          'border border-neutral-4 dark:border-neutral-8',
+          'border border-neutral-1 dark:border-neutral-9',
           {
             'bg-cream-1 dark:bg-space-3': !checked,
           },