+++ packages/tailwind-config/src/index.ts
@@ -1,6 +1,7 @@
 import RootTailwindConfig from './RootUI.tailwind.config';
 import { RootPlugins } from './plugins';
 
+export * from './types';
 export default [
   /**
    * theme is positioned first to allow other plugins to override
@@ -8,5 +9,3 @@ export default [
   RootTailwindConfig,
   ...RootPlugins,
 ];
-
-export * from './types';